Adding newsletter content to our site
Adding newsletter content to our site. Commented-out iframe contains a hard-coded URL and the height is also hard-coded; the newsletter embedded via that URL has a subscribe button/etc. at the top that may be confusing, so we need to figure out how to remove that before embedding the iframe.